Move era constants to a separate file
